+++ test/Fixture/Definition/Definitions/ImplementsDefinitionButThrowsExceptionDuringConstruction/RepositoryDefinition.php
@@ -16,6 +16,7 @@ namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsDefi
 use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
+use Faker\Generator;
 
 final class RepositoryDefinition implements Definition
 {
@@ -24,7 +25,7 @@ final class RepositoryDefinition implements Definition
         throw new \RuntimeException();
     }
 
-    public function accept(FixtureFactory $factory): void
+    public function accept(FixtureFactory $factory, Generator $faker): void
     {
         $factory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
     }