+++ test/Fixture/Definition/Definitions/ImplementsFakerAwareDefinition/RepositoryDefinition.php
@@ -16,10 +16,11 @@ namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsFake
 use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
+use Faker\Generator;
 
 final class RepositoryDefinition implements Definition
 {
-    public function accept(FixtureFactory $factory): void
+    public function accept(FixtureFactory $factory, Generator $faker): void
     {
         $factory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
     }